+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/client/render/CubeIndicatorRenderer.kt
@@ -1,6 +1,6 @@
 package dev.aaronhowser.mods.irregular_implements.client.render
 
-import dev.aaronhowser.mods.aaron.client.render.RenderUtil
+import dev.aaronhowser.mods.aaron.client.render.AaronRenderUtil
 import dev.aaronhowser.mods.aaron.misc.AaronDsls.withPose
 import net.minecraft.client.Minecraft
 import net.minecraft.core.BlockPos
@@ -55,10 +55,19 @@ object CubeIndicatorRenderer {
 			poseStack.translate(-cameraPos.x, -cameraPos.y, -cameraPos.z)
 
 			for (indicator in cubeIndicators) {
-				RenderUtil.renderCubeThroughWalls(
+				val radius = indicator.size / 2
+				val center = indicator.target.center
+				val minX = center.x - radius
+				val minY = center.y - radius
+				val minZ = center.z - radius
+				val maxX = center.x + radius
+				val maxY = center.y + radius
+				val maxZ = center.z + radius
+
+				AaronRenderUtil.renderCubeThroughWalls(
 					poseStack,
-					indicator.target.center,
-					indicator.size,
+					minX, minY, minZ,
+					maxX, maxY, maxZ,
 					indicator.color
 				)
 			}